+++ plume/src/main/kotlin/io/github/plume/oss/passes/structure/TypePass.kt
@@ -31,8 +31,6 @@ open class TypePass(private val driver: IDriver) : IProgramStructurePass {
      *
      *     TYPE_DECL -(SOURCE_FILE)-> FILE
      *     TYPE_DECL <-(CONTAINS)- FILE
-     *     TYPE_DECL -(AST)-> *MEMBER
-     *     TYPE_DECL -(AST)-> *MODIFIER
      *     TYPE_DECL -(REF)-> TYPE
      *     TYPE_DECL <-(AST)- NAMESPACE_BLOCK
      */
@@ -41,8 +39,6 @@ open class TypePass(private val driver: IDriver) : IProgramStructurePass {
             .forEach { c ->
                 logger.debug("Building type declaration, modifiers and fields for ${c.type}")
                 buildTypeDeclaration(c.type)?.let { t ->
-                    linkModifiers(c, t)
-                    linkMembers(c, t)
                     linkSourceFile(c, t)
                     linkNamespaceBlock(c, t)
                 }
@@ -61,24 +57,6 @@ open class TypePass(private val driver: IDriver) : IProgramStructurePass {
             driver.addEdge(t, f, SOURCE_FILE)
             driver.addEdge(f, t, CONTAINS)
         }
-    }
-
-    /*
-     * TYPE_DECL -(AST)-> MEMBER
-     */
-    private fun linkMembers(c: SootClass, t: NewTypeDeclBuilder) {
-        c.fields.forEachIndexed { i, field ->
-            projectMember(field, i + 1).let { memberVertex -> driver.addEdge(t, memberVertex, AST) }
-        }
-    }
-
-    /*
-     * TYPE_DECL -(AST)-> MODIFIER
-     */
-    private fun linkModifiers(c: SootClass, t: NewTypeDeclBuilder) {
-        SootParserUtil.determineModifiers(c.modifiers)
-            .mapIndexed { i, m -> NewModifierBuilder().modifierType(m).order(i + 1) }
-            .forEach { m -> driver.addEdge(t, m, AST) }
     }
 
     /**
@@ -100,13 +78,6 @@ open class TypePass(private val driver: IDriver) : IProgramStructurePass {
         driver.addEdge(t, td, REF)
         return td
     }
-
-    private fun projectMember(field: SootField, childIdx: Int): NewMemberBuilder =
-        NewMemberBuilder()
-            .name(field.name)
-            .code(field.declaration)
-            .typeFullName(field.type.toQuotedString())
-            .order(childIdx)
 
     /*
      * TYPE_DECL <-(AST)- NAMESPACE_BLOCK