+++ plume/src/main/kotlin/io/github/plume/oss/passes/update/MarkClassForRebuild.kt
@@ -1,16 +1,19 @@
-package io.github.plume.oss.passes.structure
+package io.github.plume.oss.passes.update
 
-import io.github.plume.oss.store.DriverCache
 import io.github.plume.oss.domain.mappers.VertexMapper
 import io.github.plume.oss.drivers.IDriver
+import io.github.plume.oss.passes.FileChange
 import io.github.plume.oss.passes.IProgramStructurePass
+import io.github.plume.oss.store.DriverCache
 import io.github.plume.oss.store.LocalCache
 import io.github.plume.oss.store.PlumeStorage
+import io.github.plume.oss.util.SootParserUtil
 import io.github.plume.oss.util.SootToPlumeUtil
-import io.shiftleft.codepropertygraph.generated.EdgeTypes.*
+import io.shiftleft.codepropertygraph.generated.EdgeTypes.AST
+import io.shiftleft.codepropertygraph.generated.EdgeTypes.REF
 import io.shiftleft.codepropertygraph.generated.NodeKeyNames.*
-import io.shiftleft.codepropertygraph.generated.NodeTypes.FILE
-import io.shiftleft.codepropertygraph.generated.NodeTypes.TYPE_DECL
+import io.shiftleft.codepropertygraph.generated.NodeTypes
+import io.shiftleft.codepropertygraph.generated.NodeTypes.*
 import io.shiftleft.codepropertygraph.generated.nodes.*
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
@@ -21,24 +24,23 @@ import soot.SootClass
  * Checks a list of given classes and any class which is in the database but has changed will be removed. This removal
  * involves deleting the associated FILE, METHOD, TYPE_DECL and any V where (TYPE_DECL)-AST->(v).
  */
-class MarkForRebuildPass(private val driver: IDriver) : IProgramStructurePass {
+class MarkClassForRebuild(private val driver: IDriver) {
 
-    private val logger: Logger = LogManager.getLogger(MarkForRebuildPass::javaClass)
+    private val logger: Logger = LogManager.getLogger(MarkClassForRebuild::javaClass)
     private val cache = DriverCache(driver)
 
-    private enum class FileChange { UPDATE, NEW, NOP }
-
     /**
-     * This pass will check and remove any classes that require an update from the database.
+     * This pass will check and filter classes which need updates and needed to be built. Furthermore, classes which
+     * need updates will have their modifiers rebuilt.
      *
-     * @return a list of classes which need to be built.
+     * @return a list of classes which need to be built or updated.
      */
-    override fun runPass(cs: List<SootClass>): List<SootClass> {
+    fun runPass(cs: List<SootClass>): List<Pair<SootClass, FileChange>> {
         val cStateList = cs.map(::checkIfClassNeedsAnUpdate).toList()
         val csToUpdate = cStateList.filter { it.second == FileChange.UPDATE }.toList()
-        csToUpdate.forEach { dropClassFromGraph(it.first) }
+        csToUpdate.map { it.first }.forEach(::updateModifiers)
         if (csToUpdate.isNotEmpty()) logger.info("Number of classes to update is ${csToUpdate.size}")
-        return cStateList.filterNot { it.second == FileChange.NOP }.map { it.first }.toList()
+        return cStateList.filterNot { it.second == FileChange.NOP }.toList()
     }
 
     /**
@@ -51,7 +53,7 @@ class MarkForRebuildPass(private val driver: IDriver) : IProgramStructurePass {
      */
     private fun checkIfClassNeedsAnUpdate(c: SootClass): Pair<SootClass, FileChange> {
         val newCName = SootToPlumeUtil.sootClassToFileName(c)
-        cache.tryGetFile(newCName)?.let { oldCNode: NewNodeBuilder ->
+        cache.tryGetFile(newCName)?.let { oldCNode: NewFileBuilder ->
             val currentCHash = PlumeStorage.getFileHash(c)
             logger.debug("Found an existing class with name ${c.name}")
             return if (oldCNode.build().properties().get(HASH).get() != currentCHash) {
@@ -66,16 +68,24 @@ class MarkForRebuildPass(private val driver: IDriver) : IProgramStructurePass {
         return Pair(c, FileChange.NEW)
     }
 
-    private fun dropClassFromGraph(c: SootClass) {
-        val newCName = SootToPlumeUtil.sootClassToFileName(c)
-        driver.getVerticesByProperty(NAME, newCName, FILE).filterIsInstance<NewFileBuilder>().firstOrNull()
-            ?.let { oldCNode: NewFileBuilder ->
-                driver.getNeighbours(oldCNode).use { fNeighbours ->
-                    val nodes = fNeighbours.nodes().asSequence().toList()
-                    dropTypeDecl(nodes, c.type.toQuotedString())
-                    dropFile(oldCNode)
-                }
+    private fun updateModifiers(c: SootClass) {
+        cache.tryGetTypeDecl(c.type.toQuotedString())?.let { typeDecl ->
+            // Delete old ones
+            driver.getNeighbours(typeDecl).use { g ->
+                g.nodes(MODIFIER).forEach { driver.deleteVertex(it.id(), it.label()) }
             }
+            // Create new ones
+            linkModifiers(c, typeDecl)
+        }
+    }
+
+    /*
+     * TYPE_DECL -(AST)-> MODIFIER
+     */
+    private fun linkModifiers(c: SootClass, t: NewTypeDeclBuilder) {
+        SootParserUtil.determineModifiers(c.modifiers)
+            .mapIndexed { i, m -> NewModifierBuilder().modifierType(m).order(i + 1) }
+            .forEach { m -> driver.addEdge(t, m, AST) }
     }
 
     private fun dropMethod(m: Method) {
@@ -90,7 +100,7 @@ class MarkForRebuildPass(private val driver: IDriver) : IProgramStructurePass {
                 val m2Build = m2.build()
                 driver.getNeighbours(m2).use { ns ->
                     if (ns.V(m1.id()).hasNext()) {
-                        ns.V(m1.id()).next().`in`(CALL).asSequence()
+                        ns.V(m1.id()).next().`in`(NodeTypes.CALL).asSequence()
                             .filterIsInstance<Call>()
                             .forEach {
                                 PlumeStorage.storeCallEdge(
@@ -104,36 +114,4 @@ class MarkForRebuildPass(private val driver: IDriver) : IProgramStructurePass {
         }
     }
 
-    private fun dropFile(c: NewFileBuilder) {
-        val fileName = c.build().name()
-        logger.debug("Deleting $FILE for $fileName")
-        LocalCache.removeFile(fileName)
-        driver.deleteVertex(c.id(), FILE)
-    }
-
-    private fun dropTypeDecl(ns: List<Node>, typeFullName: String) {
-        ns.filterIsInstance<TypeDecl>()
-            .filter { it.property(FULL_NAME) == typeFullName }
-            .forEach { typeDecl ->
-                logger.debug("Deleting $TYPE_DECL $typeFullName")
-                LocalCache.removeTypeDecl(typeFullName)
-                LocalCache.removeType(typeFullName)
-                driver.getNeighbours(VertexMapper.mapToVertex(typeDecl)).use { n ->
-                    n.nodes(typeDecl.id()).next()
-                        .out(AST)
-                        .forEach {
-                            logger.debug("Deleting (${it.label()}: ${it.id()})")
-                            if (it is Method) dropMethod(it)
-                            else driver.deleteVertex(it.id(), it.label())
-                        }
-                    n.nodes(typeDecl.id()).next()
-                        .`in`(REF).asSequence().distinct()
-                        .forEach {
-                            logger.debug("Deleting (${it.label()}: ${it.id()})")
-                            driver.deleteVertex(it.id(), it.label())
-                        }
-                }
-                driver.deleteVertex(typeDecl.id(), TYPE_DECL)
-            }
-    }
 }